backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pandas
openpyxl
python-multipart
requests
